fix(cli): restore module size and manpage classification budgets
Two required public smokes fail on current `main`.

`cli-command-module-size-ownership-command-modularization-smoke.py` reports
`project_lifecycle.py has 1007 lines, above budget 1000`. The file crossed the
budget when #4169 added the mutually exclusive external-sink delivery flags.
Extract the three typed inline input codecs into
`project_lifecycle_inputs.py`, which returns the owner to 925 lines without
changing any public invocation.

`cli-help-manpage-smoke.py` reports `unclassified: ['agent-context']`. The
command shipped in #4244 without a manpage class. Add it to the existing
`MANPAGE_COMMAND_HELP_ONLY` set, which is where comparable read-only
lifecycle helpers already live.

Extraction keeps the existing ownership contract: the registration and
dispatch markers asserted by
`cli-project-lifecycle-command-modularization-smoke.py` stay in the module,
and `PROJECT_LIFECYCLE_COMMANDS` still covers all four commands.

Validation:
- `python3 examples/cli-command-module-size-ownership-command-modularization-smoke.py` -> ok
- `python3 examples/cli-help-manpage-smoke.py` -> ok
- `python3 examples/cli-project-lifecycle-command-modularization-smoke.py` -> ok
- `python3 regression/cli-command-module-contract.py` -> ok
- `python -m pytest tests/cli_commands/ tests/control_plane/test_cli_output_budget.py` -> 94 passed
